+++ packages/react/package.json
@@ -39,7 +39,9 @@
     "@floating-ui/react": "^0.15.1",
     "@livekit/components-core": "0.1.13",
     "@react-hook/resize-observer": "^1.2.6",
-    "clsx": "^1.2.1"
+    "clsx": "^1.2.1",
+    "email-regex": "^5.0.0",
+    "url-regex": "^5.0.0"
   },
   "peerDependencies": {
     "react": ">=18",
@@ -54,8 +56,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tsup": "^6.5.0",
-    "typescript": "^4.9.4",
-    "url-regex": "^5.0.0"
+    "typescript": "^4.9.4"
   },
   "engines": {
     "node": ">=10"
