+++ packages/react/src/utils.ts
@@ -73,36 +73,3 @@ export const useSize = (target: React.RefObject<HTMLDivElement>) => {
   useResizeObserver(target, (entry) => setSize(entry.contentRect));
   return size;
 };
-
-export type TokenizeGrammar = { [type: string]: RegExp };
-
-export function tokenize(input: string, grammar: TokenizeGrammar) {
-  const matches = Object.entries(grammar)
-    .map(([type, rx], weight) =>
-      Array.from(input.matchAll(rx)).map(({ index, 0: content }) => ({
-        type,
-        weight,
-        content,
-        index: index ?? 0,
-      })),
-    )
-    .flat()
-    .sort((a, b) => {
-      const d = a.index - b.index;
-      return d !== 0 ? d : a.weight - b.weight;
-    })
-    .filter(({ index }, i, arr) => {
-      if (i === 0) return true;
-      const prev = arr[i - 1];
-      return prev.index + prev.content.length <= index;
-    });
-
-  const tokens = [];
-  let pos = 0;
-  for (const { type, content, index } of matches) {
-    if (index > pos) tokens.push(input.substring(pos, index));
-    tokens.push({ type, content });
-    pos = index + content.length;
-  }
-  return tokens;
-}
